// src/API/useCases/POS/Settings/entity.Settings.ts
export interface EntityPOSGeneral {
    UseOnScreenKeypad?: boolean;
    ShowFixedKeypad?: boolean;
    AutoLaunchCustomerScreen?: boolean;
    ShowTotalOnCustomerScreen?: boolean;
    ShowPriceOnCustomerScreen?: boolean;
    ShowCustomerNumber?: boolean;
    ShowSaleNotes?: boolean;
    AlwaysShowShortcuts?: boolean;
    EnableWeborderNotifications?: boolean;
    HideTotalDetails?: boolean;
    WeborderNotificationSound?: string;
    TasklistNotificationSound?: string;
    SearchFontSize?: string;
    POSFontSize?: string;
    CustomerScreenFontSize?: string;
    CustomerScreenZoom?: string;
}